requirements.txt
asgiref==3.4.1
Django==3.2.8
django-datetime-widget==0.9.3
django-filter==21.1
django-selectable==1.3.0
django-tables2==2.4.1
django-tables2-simplefilter==0.1
django-wkhtmltopdf==3.3.0
docutils==0.17.1
pytz==2021.3
six==1.16.0
South==1.0.2
sqlparse==0.4.2
wkhtmltopdf==0.2
